fix: move lifespan declaration to main
